# Only compute BitArray bit index and mask once