--- a/src/main/java/com/github/jparkie/pdd/BitArray.java
+++ b/src/main/java/com/github/jparkie/pdd/BitArray.java
@@ -38,13 +38,25 @@ public final class BitArray {
         return (int) numWords;
     }
 
+    private final class indexAndMask {
+        public final int index;
+        public final long mask;
+
+        public indexAndMask(long index) {
+            this.index = (int)(index >>> 6);
+            this.mask = 1L << (index & 0x3F);
+        }
+    }
+
     public boolean get(long index) {
-        return (data[(int) (index >>> 6)] & (1L << index)) != 0;
+        final indexAndMask iam = new indexAndMask(index);
+        return (data[iam.index] & iam.mask) != 0;
     }
 
     public boolean set(long index) {
-        if (!get(index)) {
-            data[(int) (index >>> 6)] |= (1L << index);
+        final indexAndMask iam = new indexAndMask(index);
+        if ((data[iam.index] & iam.mask) == 0) {
+            data[iam.index] |= iam.mask;
             bitCount++;
             return true;
         }
@@ -52,8 +64,9 @@ public final class BitArray {
     }
 
     public boolean clear(long index) {
-        if (get(index)) {
-            data[(int) (index >>> 6)] &= ~(1L << index);
+        final indexAndMask iam = new indexAndMask(index);
+        if ((data[iam.index] & iam.mask) != 0) {
+            data[iam.index] &= ~iam.mask;
             bitCount--;
             return true;
         }