Centralize the calculation of the array index and mask into a helper class to ensure the calculations only get done once.