@IDisposable 2016-12-07
Updated to remove the inner class (I'm new to the JVM world... the CLR JIT would have completely inlined this, I was striving to emulate the `out` parameters for C#)

Now builds the arrayIndex and bitMask inside the function and reuses inline. This still leaves the actual code duplicated, but at the level this code is to be used I think the GC pressure vs. the recompute vs. the possible copy-pasta mistakes leans to this solution... *but I bow to your experience* with Java/JVM