@jparkie 2016-12-08
Based on your GitHub handle, I had a guess you were experienced in C#. Since this class won't be touched too often and is final, I'm not too concerned about copy-paste errors.

I'll setup JMH benchmarks and check what performance gains or losses occur with the change.