@IDisposable 2016-12-08
I would love to see if it really matters, I'm barely over the hurdle with Intellij and Maven, so glad you're willing to take on that!