@IDisposable 2016-12-08
removed the shift-limiting.