@jparkie 2016-12-08
**Before**:
```
Benchmark                          Mode  Cnt          Score         Error  Units
BitArrayBenchmark.benchmarkClear  thrpt  200   22563935.037 ±  417953.613  ops/s
BitArrayBenchmark.benchmarkGet    thrpt  200  282401627.693 ± 1870596.504  ops/s
BitArrayBenchmark.benchmarkSet    thrpt  200   23075666.015 ±   58029.165  ops/s
```

**After**:
```
Benchmark                          Mode  Cnt          Score          Error  Units
BitArrayBenchmark.benchmarkClear  thrpt  200   23023640.191 ±    77683.291  ops/s
BitArrayBenchmark.benchmarkGet    thrpt  200  267436329.328 ± 11647667.023  ops/s
BitArrayBenchmark.benchmarkSet    thrpt  200   23079096.729 ±    61818.777  ops/s
```