@jparkie 2016-12-08
No. It's fine. Primarily, I don't want the performance to be worse. Although they are identical in the benchmark, in my opinion, an array implementation should be as fast as it can, and it isn't a thing that would be changed often.

Although your byte code takes more space in memory and uses more local variables, it has a smaller stack size, and you don't have the performance penalty of the first few virtual function invocation clearing their own stacks. Based on these heuristics, You win. :) I'll merge it soon.

Before Java Byte Code:
```
// access flags 0x1
  public set(J)Z
   L0
    LINENUMBER 46 L0
    ALOAD 0
    LLOAD 1
    INVOKEVIRTUAL com/github/jparkie/pdd/BitArray.get (J)Z
    IFNE L1
   L2
    LINENUMBER 47 L2
    ALOAD 0
    GETFIELD com/github/jparkie/pdd/BitArray.data : [J
    LLOAD 1
    BIPUSH 6
    LUSHR
    L2I
    DUP2
    LALOAD
    LCONST_1
    LLOAD 1
    L2I
    LSHL
    LOR
    LASTORE
   L3
    LINENUMBER 48 L3
    ALOAD 0
    DUP
    GETFIELD com/github/jparkie/pdd/BitArray.bitCount : J
    LCONST_1
    LADD
    PUTFIELD com/github/jparkie/pdd/BitArray.bitCount : J
   L4
    LINENUMBER 49 L4
    ICONST_1
    IRETURN
   L1
    LINENUMBER 51 L1
   FRAME SAME
    ICONST_0
    IRETURN
   L5
    LOCALVARIABLE this Lcom/github/jparkie/pdd/BitArray; L0 L5 0
    LOCALVARIABLE index J L0 L5 1
    MAXSTACK = 8
    MAXLOCALS = 3
```

After Java Byte Code:
```
// access flags 0x1
  public set(J)Z
   L0
    LINENUMBER 48 L0
    LLOAD 1
    BIPUSH 6
    LUSHR
    L2I
    ISTORE 3
   L1
    LINENUMBER 49 L1
    LCONST_1
    LLOAD 1
    L2I
    LSHL
    LSTORE 4
   L2
    LINENUMBER 50 L2
    ALOAD 0
    GETFIELD com/github/jparkie/pdd/BitArray.data : [J
    ILOAD 3
    LALOAD
    LLOAD 4
    LAND
    LCONST_0
    LCMP
    IFNE L3
   L4
    LINENUMBER 51 L4
    ALOAD 0
    GETFIELD com/github/jparkie/pdd/BitArray.data : [J
    ILOAD 3
    DUP2
    LALOAD
    LLOAD 4
    LOR
    LASTORE
   L5
    LINENUMBER 52 L5
    ALOAD 0
    DUP
    GETFIELD com/github/jparkie/pdd/BitArray.bitCount : J
    LCONST_1
    LADD
    PUTFIELD com/github/jparkie/pdd/BitArray.bitCount : J
   L6
    LINENUMBER 53 L6
    ICONST_1
    IRETURN
   L3
    LINENUMBER 55 L3
   FRAME APPEND [I J]
    ICONST_0
    IRETURN
   L7
    LOCALVARIABLE this Lcom/github/jparkie/pdd/BitArray; L0 L7 0
    LOCALVARIABLE index J L0 L7 1
    LOCALVARIABLE arrayIndex I L1 L7 3
    LOCALVARIABLE bitMask J L2 L7 4
    MAXSTACK = 6
    MAXLOCALS = 6
```